pyproject.toml
[project]
name = "glaze"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "aiofiles==25.1.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "asgiref==3.11.1",
    "attrs==26.1.0",
    "brotli==1.2.0",
    "certifi==2026.4.22",
    "cffi==2.0.0",
    "charset-normalizer==3.4.7",
    "click==8.3.3",
    "cloudinary==1.44.2",
    "coverage[toml]==7.13.5",
    "cryptography==46.0.7",
    "diff-match-patch==20241021",
    "dj-database-url==2.3.0",
    "django==6.0.4",
    "django-admin-sortable2==2.3.1",
    "django-cors-headers==4.9.0",
    "django-import-export==4.3.9",
    "django-meta==2.5.1",
    "django-stubs-ext==6.0.2",
    "django-stubs[compatible-mypy]==6.0.2",
    "djangorestframework==3.17.1",
    "djangorestframework-stubs[compatible-mypy]==3.16.9",
    "drf-spectacular==0.29.0",
    "fastapi==0.136.1",
    "flatbuffers==25.12.19",
    "google-auth==2.49.1",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.14",
    "imageio==2.37.3",
    "inflection==0.5.1",
    "iniconfig==2.3.0",
    "jsonnet==0.22.0",
    "jsonpath-ng==1.8.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "lazy-loader==0.5",
    "librt==0.8.1",
    "llvmlite==0.47.0",
    "logtail-python==0.3.4",
    "msgpack==1.1.2",
    "mypy==1.20.0",
    "mypy-extensions==1.1.0",
    "networkx==3.6.1",
    "numba==0.65.1",
    "numpy==2.4.4",
    "onnxruntime==1.26.0",
    "opencv-python-headless==4.13.0.92",
    "opentelemetry-exporter-otlp-proto-grpc>=1.11.1",
    "opentelemetry-instrumentation-django>=0.62b1",
    "opentelemetry-instrumentation-psycopg2>=0.62b1",
    "opentelemetry-sdk>=1.41.1",
    "packaging==26.2",
    "pathspec==1.0.4",
    "pillow==12.2.0",
    "pillow-heif==1.3.0",
    "platformdirs==4.9.6",
    "pluggy==1.6.0",
    "pooch==1.9.0",
    "protobuf==7.34.1",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.13.4",
    "pydantic-core==2.46.4",
    "pygments==2.20.0",
    "pymatting==1.1.15",
    "pytest==9.0.3",
    "pytest-cov==7.1.0",
    "pytest-django==4.10.0",
    "pytest-env==1.6.0",
    "python-dotenv==1.2.2",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "rembg[cpu]==2.0.75",
    "requests==2.33.1",
    "rpds-py==0.30.0",
    "ruff==0.12.1",
    "scikit-image==0.26.0",
    "scipy==1.17.1",
    "six==1.17.0",
    "sqlparse==0.5.5",
    "starlette==1.0.0",
    "tablib==3.9.0",
    "tifffile==2026.5.2",
    "tqdm==4.67.3",
    "types-pyyaml==6.0.12.20250915",
    "types-requests==2.33.0.20260408",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2026.2",
    "uritemplate==4.2.0",
    "urllib3==2.7.0",
    "uvicorn==0.46.0",
    "whitenoise==6.9.0",
]
